api-policy-component-service/src/test/java/com/ft/up/apipolicy/EquivalentIgnoringWindowsLineEndingsTest.java
package com.ft.up.apipolicy;

import org.junit.Test;

import static com.ft.up.apipolicy.EquivalentIgnoringWindowsLineEndings.equivalentToUnixString;
import static org.hamcrest.CoreMatchers.not;
import static org.junit.Assert.assertThat;

/**
 * EquivalentIgnoringWindowsLineEndingsTest
 *
 * @author Simon.Gibbs
 */
public class EquivalentIgnoringWindowsLineEndingsTest {

    @Test
    public void shouldMatchAWindowsStringToUNIXOne() {
        assertThat("hello\r\nworld", equivalentToUnixString("hello\nworld"));
    }

    @Test
    public void shouldNotMatchDifferentStrings() {
        assertThat("hello\r\nfred", not(equivalentToUnixString("hello\nworld")));
    }

    @Test
    public void shouldNotTreatOtherWhitespaceAsEquivalent() {
        assertThat("hello\tworld", not(equivalentToUnixString("hello world")));
    }

}
